% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNB.R
\name{CNB}
\alias{CNB}
\title{The marginal loglikelihood of the multivariate count outcome.}
\usage{
CNB(Ct, g)
}
\arguments{
\item{Ct}{A vector of categorical count.}

\item{g}{A function of expected values for each category of the multivariate count outcome.}
}
\value{
a value of marginal loglikelihood  of Dirichlet - multinomial distribution.
}
\description{
The marginal loglikelihood of the multivariate count outcome.
}
\examples{
Ct <- c(179, 12,9)
g <- c(0.65, 0.25, 0.10)
CNB(Ct,g)
}
