% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{DataGen}
\alias{DataGen}
\title{Generating Dataset consisting of Multivariate Count Outcome}
\usage{
DataGen(N, Q, S, beta, theta, s.u, n.cov, lvl.cov, Des, method, T = 2)
}
\arguments{
\item{N}{The number of subject in the generated dataset.}

\item{Q}{The number of categories in the multivariate count outcome.}

\item{S}{The total number of multivariate count outcome per subject at each time point.}

\item{theta}{The overdispersion parameter.}

\item{s.u}{The variance of random effect to account for the correlation within time point.}

\item{n.cov}{The number of covariates.}

\item{Des}{Design matrix for fixed effects.}

\item{method}{An argument of either \code{DMM} or \code{UNBM}}
}
\value{
A list of two elements where the first element is an \eqn{N\times Q} matrix of multivariate count outcome and the second element
is \code{N}$\times$ \code{n.cov} categorical covariates.
}
\description{
Generating Dataset consisting of Multivariate Count Outcome
}
\examples{
See vignette
}
