% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesMat.R
\name{DesMat}
\alias{DesMat}
\title{Design Matrix of a loglinear regression}
\usage{
DesMat(Q, lvl.cov)
}
\arguments{
\item{Q}{A number of categories in a multivariate count outcome.}

\item{lvl.cov}{A list containing the level of each covariate.}
}
\value{
A list of two elements where the first element is the left hand side of loglinear regression and the second element is the design matrix of loglinear regression.
The first element is a matrix of all possible combination of multivariate count and categorical covariate levels. See vignette for more detail.
}
\description{
Design Matrix of a loglinear regression
}
