% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF.R
\name{FF}
\alias{FF}
\title{A loglikelihood value of a subject given a random effect \code{z}}
\usage{
FF(z, u, G, theta, Cs.vec)
}
\arguments{
\item{z}{A random effect value.}

\item{u}{A standard deviation of random effect.}

\item{G}{Fixed effect regression.}

\item{theta}{An overdispersion parameter.}

\item{Cs.vec}{A vector of categorical covariate for a specific subject.}
}
\description{
A loglikelihood value of a subject given a random effect \code{z}
}
