% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixEf.R
\name{FixEf}
\alias{FixEf}
\title{The Fixed Effect of a Regression on Count Data.}
\usage{
FixEf(x.vec, Q, lvl.cov, b, Des)
}
\arguments{
\item{x.vec}{The observed vector of covariate for each category.}

\item{Q}{The number of categories in a multivariate count outcome.}

\item{lvl.cov}{A list containing the level of each covariate.}

\item{b}{A vector containing estimates of covariate effects.}

\item{Des}{A list containing two matrices, the first element is the left hand side of the equation. The second element is the design matrix.
Could be a result from \code{DesMat}.}
}
\value{
A \code{Q} dimensional vector consisting of the expected values of \code{Q} categories given the \code{x.vec}.
}
\description{
The Fixed Effect of a Regression on Count Data.
}
