% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logliklong.R
\name{logliklong}
\alias{logliklong}
\title{The Loglikelihood Function for Overdispersed Multivariate Count Outcome in Repeated Measurements}
\usage{
logliklong(pars, dataset, Des, model)
}
\arguments{
\item{pars}{The vector of fixed effect parameters followed by logarithm of \code{s.u} and logarithm of the overdispersion parameter \code{theta}.}

\item{dataset}{A list of two elements, where the first element is the multivariate outcome and the second element is the matrix of categorical covariates.}

\item{Des}{A list of two elements, where the first element is the left hand side of the matrix equation of the loglinear model and the second element is the design matrix for the regression covariates.}

\item{model}{A model for the multivariate count outcome. Either \code{DMM}, \code{CNBM} or \code{UNBM}.}
}
\value{
The estimated regression parameters of regression model, the loglikelihood values, and the Hessian matrix of the estimated parameters.
This loglikelihood function is used to estimate parameters of the regression of multivariate count outcome on categorical covariate in the repeated measurement setting
in which the correlation between observation at two different time points is accounted ($\sigma^2_u \neq 0$).
}
\description{
The Loglikelihood Function for Overdispersed Multivariate Count Outcome in Repeated Measurements
}
